fix(auth): resolve redirect loop by consolidating middleware
